Remove all pairs exect VTHO and update styles
