+++ package.json
@@ -17,6 +17,7 @@
     "normalize.css": "^8.0.1",
     "qrcode": "^1.4.4",
     "react": "^16.13.1",
+    "react-copy-to-clipboard": "^5.0.4",
     "react-device-detect": "^1.17.0",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
