--- src/components/Container/index.jsx
@@ -1,1 +1,0 @@
-export { default } from './Container';
