--- src/components/NoConnex/index.jsx
@@ -1,1 +1,0 @@
-export { default } from './NoConnex';
