--- src/constants/tokens.json
@@ -1,38 +1,0 @@
-{
-  "VTHO": {
-    "symbol": "VTHO",
-    "token": "0x0000000000000000000000000000456e65726779",
-    "address": "0x012345403c589a51b02ee27bd41339f6114aac6a",
-    "exchange": "0xf9F99f982f3Ea9020f0A0afd4D4679dFEe1B63cf"
-  },
-  "OCE": {
-    "symbol": "OCE",
-    "token": "0x0CE6661b4ba86a0EA7cA2Bd86a0De87b0B860F14",
-    "address": "0x114E1C8CA8f63B7C5692daFC8CFeB4112e9Ff7b4",
-    "exchange": "0xDC391a5dbB89a3F768c41Cfa0e85dcaAF3A91f91"
-  },
-  "PLA": {
-    "symbol": "PLA",
-    "token": "0x89827F7bB951Fd8A56f8eF13C5bFEE38522F2E1F",
-    "address": "0x84497d7E46F671092f3AC8455f0F7F30890Df5A1",
-    "exchange": "0xD293f479254D5F6494c66A4982C7cA514A53D7C4"
-  },
-  "DBET": {
-    "symbol": "DBET",
-    "token": "0x1b8ec6c2a45cca481da6f243df0d7a5744afc1f8",
-    "address": "0xD18e5F2cd1012b28fB0007A578089EC6fFB6e2a3",
-    "exchange": "0x18C2385481cDf28779aC271272398dD61cc8CF3E"
-  },
-  "SHA": {
-    "symbol": "SHA",
-    "token": "0x5db3C8A942333f6468176a870dB36eEf120a34DC",
-    "address": "0x39df9f3aE0FaAfeF6B12330930900707aca7fF1A",
-    "exchange": "0xC19cf5Dfb71374b920F786078D37b5225CFcF30E"
-  },
-  "EHRT": {
-    "symbol": "EHrT",
-    "token": "0xf8e1faa0367298b55f57ed17f7a2ff3f5f1d1628",
-    "address": "0xF475B9CfB14662E42286699b519cA8f1FDA9F6a9",
-    "exchange": "0x6D08D19Dff533050f93EaAa0a009e2771D3598bC"
-  }
-}
