+++ src/fonts.css
@@ -1,0 +1,4 @@
+@font-face {
+  font-family: "VCR";
+  src: local("VCR"), url(/fonts/VCR_OSD_MONO.ttf) format("truetype");
+}
